refactor: re-inline useEnterAccountRecoveryFromMain hook
Delete the useEnterAccountRecoveryFromMain hook and its test — the hook
wraps a 2-line operation used in exactly one place. Inline the gotoSetup
and gotoPage calls directly in UpdatePasswordScreen.tsx.
