+++ frontend/components/SettingsPage/UpdatePassword/UpdatePasswordScreen.tsx
@@ -12,11 +12,11 @@ import {
   RequiredMark,
 } from '@/components/ui';
 import { PasswordStrength } from '@/components/ui/forms';
-import { COLOR } from '@/constants';
+import { COLOR, PAGES, SETUP_SCREEN } from '@/constants';
 import { ERROR_CODE } from '@/constants/errors';
 import { SettingsScreenMap } from '@/constants/screen';
 import { useMessageApi } from '@/context/MessageProvider';
-import { useEnterAccountRecoveryFromMain, useSettings } from '@/hooks';
+import { usePageState, useSettings, useSetup } from '@/hooks';
 import { AccountService } from '@/service/Account';
 
 const { Title, Text } = Typography;
@@ -30,7 +30,8 @@ type UpdatePasswordFormValues = {
 export const UpdatePasswordScreen = () => {
   const [form] = Form.useForm<UpdatePasswordFormValues>();
   const { goto } = useSettings();
-  const enterAccountRecovery = useEnterAccountRecoveryFromMain();
+  const { goto: gotoSetup } = useSetup();
+  const { goto: gotoPage } = usePageState();
   const messageApi = useMessageApi();
   const [isSubmitting, setIsSubmitting] = useState(false);
 
@@ -119,7 +120,10 @@ export const UpdatePasswordScreen = () => {
                   type="link"
                   className="text-sm w-fit p-0"
                   style={{ height: 'auto' }}
-                  onClick={enterAccountRecovery}
+                  onClick={() => {
+                    gotoSetup(SETUP_SCREEN.AccountRecovery);
+                    gotoPage(PAGES.Setup);
+                  }}
                 >
                   Forgot your password?
                 </Button>
