+++ frontend/tests/components/SettingsPage/UpdatePassword/UpdatePasswordScreen.test.tsx
@@ -2,6 +2,7 @@ import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import { act } from 'react';
 
 import { UpdatePasswordScreen } from '../../../../components/SettingsPage/UpdatePassword';
+import { PAGES, SETUP_SCREEN } from '../../../../constants';
 import { ERROR_CODE } from '../../../../constants/errors';
 import { SettingsScreenMap } from '../../../../constants/screen';
 
@@ -10,7 +11,8 @@ import { SettingsScreenMap } from '../../../../constants/screen';
 // ---------------------------------------------------------------------------
 
 const mockGoto = jest.fn();
-const mockEnterAccountRecovery = jest.fn();
+const mockSetupGoto = jest.fn();
+const mockPageGoto = jest.fn();
 const mockSuccess = jest.fn();
 const mockError = jest.fn();
 
@@ -19,7 +21,8 @@ jest.mock('../../../../hooks', () => ({
     screen: SettingsScreenMap.UpdatePassword,
     goto: mockGoto,
   }),
-  useEnterAccountRecoveryFromMain: () => mockEnterAccountRecovery,
+  useSetup: () => ({ goto: mockSetupGoto }),
+  usePageState: () => ({ goto: mockPageGoto }),
 }));
 
 jest.mock('../../../../context/MessageProvider', () => ({
@@ -256,7 +259,8 @@ describe('UpdatePasswordScreen', () => {
     it('navigates to AccountRecovery on "Forgot your password?" click', () => {
       render(<UpdatePasswordScreen />);
       fireEvent.click(screen.getByText('Forgot your password?'));
-      expect(mockEnterAccountRecovery).toHaveBeenCalled();
+      expect(mockSetupGoto).toHaveBeenCalledWith(SETUP_SCREEN.AccountRecovery);
+      expect(mockPageGoto).toHaveBeenCalledWith(PAGES.Setup);
     });
   });
 });
